Removed dead code from utils module #345 (#362)
