+++ utils/__init__.py
@@ -1,24 +1,12 @@
 from .utils import (
-    check_fragments_size,
     check_support,
-    cod,
     files_names_from_pattern,
     import_pandas_into_module_namespace,
     load_data_pandas,
     load_data_modin_on_hdk,
-    mse,
-    print_times,
     split,
     print_results,
-    convert_units,
-    write_to_csv_by_chunks,
-    create_dir,
-    make_chk,
     memory_usage,
-    join_to_tbls,
-    get_tmp_filepath,
-    FilesCombiner,
-    get_dir_size,
     getsize,
     run_benchmarks,
 )
