+++ utils_base_env/__init__.py
@@ -1,7 +1,5 @@
 from .utils_base_env import (
     execute_process,
-    find_free_port,
-    KeyValueListParser,
     str_arg_to_bool,
     add_sql_arguments,
     prepare_parser,
